backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.2
openai==1.12.0
gTTS==2.5.0
Jinja2==3.1.3
python-multipart==0.0.6
